Fix Little Endian problems and make things easier
First off, it has been made clear to me that Registers should not be
stored as Little Endian. This is simply a memory read/write order thing.

Methods Register::{bc, af, de, hl} now return bytes in the order they
are stored (which is MSB to the left).

To make this easier we:
    - Added write_u16 and read_u16 methods to the Mmu
        * These swap the bytes so that we don't have to deal with it.
    - Added push & pop stack methods to the Cpu that handle the
      incrementing, decrementing, and use Mmu methods to use the correct
      byte order.

Existing operations were also updated to use these new methods.